unified address formatting
